Added Playlist CRUD
Added Playlist CRUD functionality (add song to playlist, remove song from playlist)
Fixed "CurrentSong" component position in "Library" Screen
Improved and added backend endpoints

+++ backend/main.py
@@ -1,7 +1,8 @@
 import datetime
+import json
 import os
-from random import seed
-from typing import List, Optional
+import socket
+from typing import Dict, List, Optional
 
 import eyed3  # type: ignore
 import uvicorn  # type: ignore
@@ -13,6 +14,7 @@ from sqlmodel import Session, select
 from database_setup import create_db_and_tables, engine
 from models import (
     Liked,
+    LikedBase,
     Playlist,
     PlaylistBase,
     PlaylistRead,
@@ -21,6 +23,7 @@ from models import (
     SongBase,
     SongPlaylistLink,
     SongRead,
+    SongReadWithLike,
     SongReadWithPlaylists,
     SongUpdate,
 )
@@ -68,11 +71,18 @@ async def stream_song(song_id: int) -> StreamingResponse:
     return StreamingResponse(iterfile(), media_type="audio/mp3")
 
 
-@app.post("/songs/{song_id}/like")
+@app.post("/songs/{song_id}/like", response_model=Dict)
 async def like(song_id: int) -> dict:
+    song = session.get(Song, song_id)
+    if not song:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Song not found"
+        )
     new_like = Liked(song_id=song_id)
-    new_link = SongPlaylistLink(playlist_id=1, song_id=song_id)
+    song.liked_bool = new_like
     session.add(new_like)
+    session.add(song)
+    new_link = SongPlaylistLink(playlist_id=1, song_id=song_id)
     session.add(new_link)
     session.commit()
     return {"Liked song with id: ": song_id}
@@ -80,19 +90,32 @@ async def like(song_id: int) -> dict:
 
 @app.delete("/songs/{song_id}/unlike")
 async def unlike(song_id: int) -> dict:
-    statement = select(Liked).where(Liked.song_id == song_id)
-    stmt = select(SongPlaylistLink).where(
+    likedStmt = select(Liked).where(Liked.song_id == song_id)
+    songplaylistlinkStmt = select(SongPlaylistLink).where(
         SongPlaylistLink.song_id == song_id, SongPlaylistLink.playlist_id == 1
     )
-    liked = session.exec(statement).one_or_none()
-    link = session.exec(stmt).one_or_none()
+    song = session.get(Song, song_id)
+    if not song:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Song entry for song not found",
+        )
+    liked = session.exec(likedStmt).one_or_none()
     if not liked:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Liked entry for song not found",
         )
+    link = session.exec(songplaylistlinkStmt).one_or_none()
+    if not link:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Link entry for song not found",
+        )
+
     session.delete(liked)
     session.delete(link)
+    song.liked_bool = None
     session.commit()
     return {"Unliked song with id: ": song_id}
 
@@ -103,7 +126,6 @@ async def check(
     song_id2: int,
     song_id3: int,
     song_id4: int,
-    # song_id5: int,
     playlist_id1: int,
     playlist_id2: int,
 ):
@@ -111,7 +133,6 @@ async def check(
     song2 = session.exec(select(Song).where(Song.id == song_id2)).one()
     song3 = session.exec(select(Song).where(Song.id == song_id3)).one()
     song4 = session.exec(select(Song).where(Song.id == song_id4)).one()
-    # song5 = session.exec(select(Song).where(Song.id == song_id5)).one()
 
     playlist1 = session.exec(select(Playlist).where(Playlist.id == playlist_id1)).one()
     playlist2 = session.exec(select(Playlist).where(Playlist.id == playlist_id2)).one()
@@ -120,7 +141,6 @@ async def check(
     playlist2.songs.append(song2)
     playlist2.songs.append(song3)
     playlist2.songs.append(song4)
-    # playlist2.songs.append(song5)
     session.add(playlist1)
     session.add(playlist2)
     session.commit()
@@ -137,17 +157,64 @@ async def create_playlist(playlist: PlaylistBase) -> Playlist:
     return new_playlist
 
 
-@app.get("/songs/{playlist_id}/fetch", response_model=List)
-async def get_playlist(playlist_id: int) -> List:
-    statement = (
-        select(Song, Liked.song_id)
-        .select_from(Song)
-        .join(SongPlaylistLink)
-        .join(Playlist)
-        .join(Liked, isouter=True)
-        .where(Playlist.id == playlist_id)
-    )
-    return session.exec(statement).all()
+@app.delete(
+    "/songs/delete_playlist/{playlist_id}", status_code=status.HTTP_204_NO_CONTENT
+)
+async def delete_playlist(playlist_id: int) -> Optional[Playlist]:
+    playlist = session.get(Playlist, playlist_id)
+    if not playlist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Playlist not found"
+        )
+    session.delete(playlist)
+    session.commit()
+    return playlist
+
+
+@app.get("/songs/{playlist_id}/fetch", response_model=List[SongReadWithLike])
+async def get_playlist(playlist_id: int):
+    playlist = session.get(Playlist, playlist_id)
+    if not playlist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Playlist not found"
+        )
+    return playlist.songs
+
+
+@app.patch("/songs/add_song/{playlist_id}/{song_id}", response_model=Dict)
+async def add_song_to_playlist(song_id: int, playlist_id: int) -> dict:
+    playlist = session.get(Playlist, playlist_id)
+    if not playlist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Playlist not found"
+        )
+    song = session.get(Song, song_id)
+    if not song:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Song not found"
+        )
+    playlist.songs.append(song)
+    session.add(playlist)
+    session.commit()
+    return {"Added Song with ID: ": song_id, "to Playlist with ID: ": playlist_id}
+
+
+@app.patch("/songs/remove_song/{playlist_id}/{song_id}", response_model=Dict)
+async def delete_song_from_playlist(song_id: int, playlist_id: int) -> dict:
+    playlist = session.get(Playlist, playlist_id)
+    if not playlist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Playlist not found"
+        )
+    song = session.get(Song, song_id)
+    if not song:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Song not found"
+        )
+    playlist.songs.remove(song)
+    session.add(song)
+    session.commit()
+    return {"Deleted Song with ID: ": song_id, "From Playlist with ID: ": playlist_id}
 
 
 @app.get("/songs/playlists", response_model=List[PlaylistReadWithSongs])
@@ -155,7 +222,7 @@ async def get_all_playlist():
     return session.exec(select(Playlist)).all()
 
 
-@app.get("/songs", response_model=List[SongReadWithPlaylists])
+@app.get("/songs", response_model=List[SongReadWithLike])
 async def get_all_songs():
     return session.exec(select(Song)).all()
 
@@ -163,8 +230,6 @@ async def get_all_songs():
 @app.get("/songs/liked", response_model=List[PlaylistReadWithSongs])
 async def get_liked_songs():
     return session.exec(select(Playlist).where(Playlist.id == 1)).all()
-    # statement = select(Song, Liked.song_id).join(Liked)
-    # return session.exec(statement).all()
 
 
 @app.get("/songs/{song_id}", response_model=SongRead)
@@ -214,8 +279,21 @@ async def delete_song(song_id: int) -> Optional[Song]:
 
 
 @app.on_event("startup")
+def create_initial_playlists():
+
+    # Create Liked Songs Playlist
+    stmt = select(Playlist).where(Playlist.id == 2)
+    db_liked_playlist = session.exec(stmt).one_or_none()
+    if not db_liked_playlist:
+        session.add(Playlist(name="Liked Songs"))
+    session.commit()
+
+
+@app.on_event("startup")
 # @repeat_every(seconds=5)
 def scan_songs():
+
+    # Scan All Songs
     for song in os.listdir(music_folder_url):
         if song.endswith(".mp3"):
             joined_path = os.path.join(music_folder_url, song)
@@ -244,15 +322,6 @@ def scan_songs():
                 image_file.close()
                 artwork_exists = True
 
-            new_song = Song(
-                url=joined_path,
-                title=song_title,
-                artist=audiofile.tag.artist,
-                artwork=artwork_path if artwork_exists else None,
-                last_modify=get_last_modify(joined_path),
-                duration=audiofile.info.time_secs,
-            )
-            print(new_song.title, " added!")
             session.add(
                 Song(
                     url=joined_path,
@@ -263,13 +332,7 @@ def scan_songs():
                     duration=audiofile.info.time_secs,
                 )
             )
-    stmt = select(Playlist).where(Playlist.id == 1)
-    db_liked_playlist = session.exec(stmt).one_or_none()
-    if not db_liked_playlist:
-        session.add(
-            Playlist(name="Liked Songs")
-        )  # TODO: Think of a better way to implement
-    session.commit()
+            print(song_title + " Added to DB!")
 
 
 if __name__ == "__main__":
